requirements.txt
numpy>=1.24
scikit-learn>=1.2
requests>=2.31

# clustering + pipeline deps
hdbscan>=0.8.36
pandas>=1.5
python-dateutil>=2.8.2
beautifulsoup4>=4.12
emoji>=2.10.0
jieba>=0.42.1
tqdm>=4.66
python-dotenv>=1.0.1

# optional local embedding backend
sentence-transformers>=2.2.2
